Build weather component and remove api dependency
All other components that don't depend on the weather api should be
able to load even if the weather api data never get's sent to the client

[Closes #87]
